Multi-Docs Chatbot/pyproject.toml
[project]
name = "multi-docs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=1.4.0",
    "docx2txt>=0.9",
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.3",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-groq>=1.1.1",
    "langchain-huggingface>=1.2.0",
    "langchain-text-splitters>=1.1.0",
    "openai>=2.15.0",
    "pypdf>=6.6.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.52.2",
    "streamlit-chat>=0.1.1",
    "tiktoken>=0.12.0",
]
